added astar and removed object files

+++ heuristic.h
@@ -10,6 +10,6 @@
  *  @param y2 B.y
  *  @return the distance
  */
-double mdist(int x1, int y1, int x2, int y2);
+int mdist(int x1, int y1, int x2, int y2);
 
 #endif
